Migrate TrafficCopResult's `extra_`'s uint32_t variant to a uint64_t
For the many people who are surely returning over four-and-a-half billion rows using Terrier, there's a type-mismatch that can lead to overflow. NumRows is a 64-bit datatype, but the variant that often expresses it it is only a 32-bit datatype. For everyone else, you might one day want to compile the project with g++-10 or later. The changes in this commit (among others) are required in order for terrier to build. The ugly ull prefix on zeroes is necessary, as you'll otherwise get an error where gcc believes you're attempting to pass a character. It doesn't seem to want to do type conversion.